requirements.txt
uvicorn 
sqlalchemy 
fastapi
psycopg2
werkzeug
python-jose
python-multipart